refactor: split repository normalized review queries
